requirements.txt
google-genai
fastapi
uvicorn
pydantic
pymongo
python-dotenv
pypdf
numpy
reportlab
